v2.0.59

+++ phi/agent/file.py
@@ -1,36 +1,34 @@
-import runpy
-import functools
 from pathlib import Path
-from typing import Optional
+from typing import Optional, List
 
 from phi.agent import Agent
 from phi.utils.log import logger
 
 
-@functools.lru_cache(maxsize=None)
-def warn() -> None:
-    logger.warning("FileAgent can execute arbitrary code. Do not use without human supervision.")
-
-
 class FileAgent(Agent):
     def __init__(
-        self, base_dir: Optional[Path] = None, save_file: bool = True, run_file: bool = True, read_file: bool = True
+        self,
+        base_dir: Optional[Path] = None,
+        save_files: bool = True,
+        read_files: bool = True,
+        list_files: bool = False,
     ):
         super().__init__(name="file_agent")
 
         self.base_dir: Path = base_dir or Path.cwd()
-        if save_file:
-            self.register(self.save_contents_to_file, sanitize_arguments=False)
-        if run_file:
-            self.register(self.run_python_file)
-        if read_file:
-            self.register(self.read_contents_from_file)
+        if save_files:
+            self.register(self.save_file, sanitize_arguments=False)
+        if read_files:
+            self.register(self.read_file)
+        if list_files:
+            self.register(self.list_files)
 
-    def save_contents_to_file(self, contents: str, file_name: str) -> str:
+    def save_file(self, contents: str, file_name: str, overwrite: bool = True) -> str:
         """Saves the contents to a file called `file_name` and returns the file name if successful.
 
         :param contents: The contents to save.
         :param file_name: The name of the file to save to.
+        :param overwrite: Overwrite the file if it already exists.
         :return: The file name if successful, otherwise returns an error message.
         """
         try:
@@ -38,6 +36,8 @@ class FileAgent(Agent):
             logger.debug(f"Saving contents to {file_path}")
             if not file_path.parent.exists():
                 file_path.parent.mkdir(parents=True, exist_ok=True)
+            if file_path.exists() and not overwrite:
+                return f"File {file_name} already exists"
             file_path.write_text(contents)
             logger.info(f"Saved: {file_path}")
             return str(file_name)
@@ -45,27 +45,7 @@ class FileAgent(Agent):
             logger.error(f"Error saving to file: {e}")
             return f"Error saving to file: {e}"
 
-    def run_python_file(self, file_name: str, variable_to_return: str) -> str:
-        """Runs the Python file `file_name` and returns the value of `variable_to_return` if successful.
-
-        :param file_name: The name of the file to run.
-        :param variable_to_return: The variable to return.
-        :return: value of `variable_to_return` if successful, otherwise returns an error message.
-        """
-        try:
-            file_path = self.base_dir.joinpath(file_name)
-            logger.info(f"Running file: {file_path}")
-            globals_after_run = runpy.run_path(str(file_path), run_name="__main__")
-            result = globals_after_run.get(variable_to_return)
-            if result is None:
-                return f"Variable {variable_to_return} not found"
-            logger.debug(f"Result: {result}")
-            return str(result)
-        except Exception as e:
-            logger.error(f"Error running python file: {e}")
-            return f"Error running python file: {e}"
-
-    def read_contents_from_file(self, file_name: str) -> str:
+    def read_file(self, file_name: str) -> str:
         """Reads the contents of the file `file_name` and returns the contents if successful.
 
         :param file_name: The name of the file to read.
@@ -79,3 +59,15 @@ class FileAgent(Agent):
         except Exception as e:
             logger.error(f"Error reading file: {e}")
             return f"Error reading file: {e}"
+
+    def list_files(self) -> List[str]:
+        """Returns a list of files in the base directory
+
+        :return: The contents of the file if successful, otherwise returns an error message.
+        """
+        try:
+            logger.info(f"Reading files in : {self.base_dir}")
+            return [str(file_path) for file_path in self.base_dir.iterdir()]
+        except Exception as e:
+            logger.error(f"Error reading files: {e}")
+            return [f"Error reading files: {e}"]
